--- a/build_requirements.txt
+++ b/build_requirements.txt
@@ -3,3 +3,4 @@ Cython>=3.0
 wheel==0.38.1
 ninja
 spin==0.5
+build